Fixes #2 by adding Windows Path Support
We also added some additonal UI elements to the first run modal.
You may now cancel the modal and allow Brackets to continue to load
without setting the Base Path.

There is also a "Browse" button that will show a GUI dialog for selecting
the folder which should make things a little easier.

NOTE: This was tested on Windows 8.1 with the Brackets 1.2 Pre Release